requirements.txt
python-dotenv
pip-system-certs
psycopg2-binary
transformers 
torch 
scikit-learn
matplotlib
